#60 - refactor: Add update contract to rewards repository port
